+++ CHANGES.md
@@ -131,6 +131,7 @@
     -   Fix wrong italic highlighting in HTML attributes. ([GH-410][])
     -   Fix markdown-follow-thing-at-point issue for continuous links. ([GH-305][])
     -   Fix wrong setting major-mode issue at following wiki link([GH-427][])
+    -   Fix not consider `markdown-list-indent-width` issue([GH-405][])
 
   [gh-349]: https://github.com/jrblevin/markdown-mode/issues/349]
   [gh-171]: https://github.com/jrblevin/markdown-mode/issues/171
@@ -184,8 +185,9 @@
   [gh-350]: https://github.com/jrblevin/markdown-mode/pull/350
   [gh-369]: https://github.com/jrblevin/markdown-mode/pull/369
   [gh-378]: https://github.com/jrblevin/markdown-mode/pull/378
+  [gh-405]: https://github.com/jrblevin/markdown-mode/issues/405
   [gh-410]: https://github.com/jrblevin/markdown-mode/issues/410
-  [gh-410]: https://github.com/jrblevin/markdown-mode/issues/427
+  [gh-427]: https://github.com/jrblevin/markdown-mode/issues/427
   [gh-437]: https://github.com/jrblevin/markdown-mode/issues/437
 
 # Markdown Mode 2.3